@@ -21,11 +21,11 @@
         }
     ],
     "require": {
-        "php": "^7.2",
-        "spatie/packagist-api": "^1.2"
+        "php": "^8.0",
+        "spatie/packagist-api": "^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0"
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
         "psr-4": {
