@@ -3,21 +3,20 @@
 namespace ahinkle\PackagistLatestVersion\Tests;
 
 use ahinkle\PackagistLatestVersion\PackagistLatestVersion;
-use GuzzleHttp\Client;
 use PHPUnit\Framework\TestCase;
 
 class PackagistLatestVersionTest extends TestCase
 {
-    /** @var ahinkle\PackagistLatestVersion\PackagistLatestVersion */
-    protected $packagist;
+    /**
+     * @var PackagistLatestVersion
+     */
+    protected PackagistLatestVersion $packagist;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $client = new Client();
-
-        $this->packagist = new PackagistLatestVersion($client);
-
         parent::setUp();
+
+        $this->packagist = new PackagistLatestVersion();
     }
 
     public function test_it_can_instantiate_package_latest_version()
